rename package, simplify scripts
