+++ .gitignore
@@ -11,6 +11,4 @@ dist
 examples/**/yarn.lock
 package-lock.json
 tsconfig.tsbuildinfo
-**/*.yalc
-yalc.lock
 coverage
