+++ package.json
@@ -4,28 +4,26 @@
   "private": true,
   "description": "React Hooks library for remote data fetching",
   "workspaces": [
-    "./swr",
-    "./swr/infinite",
-    "./swr/immutable"
+    "./package"
   ],
   "scripts": {
-    "clean": "rimraf dist infinite/dist immutable/dist",
+    "clean": "rimraf package/dist package/infinite/dist package/immutable/dist",
     "build": "yarn build:core && yarn build:infinite && yarn build:immutable",
-    "watch": "yalc publish && node scripts/watch.js",
-    "watch:core": "node scripts/watch.js core",
-    "watch:infinite": "node scripts/watch.js infinite",
-    "watch:immutable": "node scripts/watch.js immutable",
-    "build:core": "bunchee core/index.ts --cwd swr --no-sourcemap",
-    "build:infinite": "bunchee index.ts --cwd swr/infinite --no-sourcemap",
-    "build:immutable": "bunchee index.ts --cwd swr/immutable --no-sourcemap",
-    "prepublishOnly": "cd swr && yarn clean && yarn build",
-    "publish-beta": "yarn publish --tag beta",
+    "prepublishOnly": "yarn clean && yarn build",
+    "build:core": "bunchee core/index.ts --cwd package --no-sourcemap",
+    "build:infinite": "bunchee index.ts --cwd package/infinite --no-sourcemap",
+    "build:immutable": "bunchee index.ts --cwd package/immutable --no-sourcemap",
+    "watch": "yarn npm-run-all -p watch:core watch:infinite watch:immutable",
+    "watch:core": "yarn build:core -w",
+    "watch:infinite": "yarn build:infinite -w",
+    "watch:immutable": "yarn build:immutable -w",
+    "publish": "yarn workspace swr publish",
+    "publish-beta": "yarn workspace swr publish --tag beta",
     "types:check": "tsc --noEmit --project tsconfig.check.json && tsc --noEmit -p test",
     "format": "prettier --write ./**/*.{ts,tsx}",
     "lint": "eslint . --ext .ts,.tsx --cache",
     "lint:fix": "yarn lint --fix",
-    "test": "jest --coverage",
-    "register": "yarn link && cd node_modules/react && yarn link"
+    "test": "jest --coverage"
   },
   "husky": {
     "hooks": {
@@ -69,7 +67,6 @@
     "react-dom-experimental": "npm:react-dom@alpha",
     "react-experimental": "npm:react@alpha",
     "rimraf": "3.0.2",
-    "typescript": "4.4.3",
-    "yalc": "1.0.0-pre.53"
+    "typescript": "4.4.3"
   }
 }
