+++ package/core/index.ts
