+++ package/core/utils/config.ts
