+++ package/core/utils/global-state.ts
