+++ package/core/utils/hash.ts
