+++ package/core/utils/merge-config.ts
