+++ package/core/utils/state.ts
