+++ package/infinite/index.ts
