+++ package/infinite/package.json
