+++ package/infinite/tsconfig.json
