+++ test/unit/utils.test.tsx
@@ -1,7 +1,7 @@
-import { normalize } from '../../swr/core/utils/normalize-args'
-import { stableHash as hash } from '../../swr/core/utils/hash'
-import { serialize } from '../../swr/core/utils/serialize'
-import { mergeConfigs } from '../../swr/core/utils/merge-config'
+import { normalize } from '../../package/core/utils/normalize-args'
+import { stableHash as hash } from '../../package/core/utils/hash'
+import { serialize } from '../../package/core/utils/serialize'
+import { mergeConfigs } from '../../package/core/utils/merge-config'
 
 describe('Utils', () => {
   it('should normalize arguments correctly', async () => {
