+++ test/use-swr-local-mutation.test.tsx
@@ -1,7 +1,7 @@
 import { act, screen, fireEvent } from '@testing-library/react'
 import React, { useEffect, useState } from 'react'
 import useSWR, { mutate as globalMutate, useSWRConfig } from 'swr'
-import { serialize } from '../swr/core/utils/serialize'
+import { serialize } from '../package/core/utils/serialize'
 import {
   createResponse,
   sleep,
