+++ test/use-swr-middlewares.test.tsx
@@ -1,7 +1,7 @@
 import { act, screen } from '@testing-library/react'
 import React, { useState, useEffect, useRef } from 'react'
 import useSWR, { Middleware, SWRConfig } from 'swr'
-import { withMiddleware } from '../swr/core/utils/with-middleware'
+import { withMiddleware } from '../package/core/utils/with-middleware'
 
 import {
   createResponse,
