+++ tsconfig.check.json
@@ -3,5 +3,5 @@
   "compilerOptions": {
     "rootDir": ".",
   },
-  "include": ["swr/**/*",]
+  "include": ["package/**/*",]
 }
